Fix broken loading state handling
