+++ src/components/PortfolioChart.tsx
@@ -1,13 +1,15 @@
-import type { Coin } from "@/types/Coin";
 import { Card, DonutChart, Legend, Title } from "@tremor/react";
 import { formatCurrency } from "@/utils/formatCurrency";
 
 type PortfolioChartProps = {
-  ownedCoins: Coin[];
+  coins: {
+    name: string;
+    totalValue: number;
+  }[];
 };
 
-const PortfolioChart = ({ ownedCoins }: PortfolioChartProps) => {
-  if (ownedCoins.length === 0)
+const PortfolioChart = ({ coins }: PortfolioChartProps) => {
+  if (coins.length === 0)
     return (
       <Card>
         <div className="flex h-96 flex-col items-center justify-center">
@@ -20,16 +22,13 @@ const PortfolioChart = ({ ownedCoins }: PortfolioChartProps) => {
     <Card hFull={true}>
       <Title>Portfolio</Title>
       <DonutChart
-        data={ownedCoins}
+        data={coins}
         category="totalValue"
         dataKey="name"
         valueFormatter={formatCurrency}
         marginTop="mt-6"
       />
-      <Legend
-        categories={ownedCoins.map((coin) => coin.name)}
-        marginTop="mt-6"
-      />
+      <Legend categories={coins.map((coin) => coin.name)} marginTop="mt-6" />
     </Card>
   );
 };
