--- src/types/Coin.ts
@@ -1,8 +1,0 @@
-export type Coin = {
-  symbol: string;
-  quantity: number;
-  name: string;
-  currentPrice: number;
-  totalValue: number;
-  percentChange24h: number;
-};
