--- src/types/MarketCoin.ts
@@ -1,10 +1,0 @@
-export type MarketCoin = {
-  name: string;
-  symbol: string;
-  rank: number;
-  price: number;
-  percentChange1h: number;
-  percentChange24h: number;
-  percentChange7d: number;
-  marketCap: number;
-};
